l1b_lambda/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile --resolver=backtracking
#
attrs==22.1.0
    # via
    #   cattrs
    #   jsii
aws-cdk-asset-awscli-v1==2.2.30
    # via aws-cdk-lib
aws-cdk-asset-kubectl-v20==2.1.1
    # via aws-cdk-lib
aws-cdk-asset-node-proxy-agent-v5==2.0.38
    # via aws-cdk-lib
aws-cdk-aws-lambda-python-alpha==2.54.0a0
    # via -r requirements.in
aws-cdk-lib==2.54.0
    # via
    #   -r requirements.in
    #   aws-cdk-aws-lambda-python-alpha
boto3==1.26.29
    # via -r requirements.in
boto3-stubs==1.26.29
    # via types-boto3
botocore==1.29.29
    # via
    #   boto3
    #   s3transfer
botocore-stubs==1.29.29
    # via
    #   boto3-stubs
    #   types-botocore
build==0.9.0
    # via -r requirements.in
cattrs==22.2.0
    # via jsii
constructs==10.1.189
    # via
    #   aws-cdk-aws-lambda-python-alpha
    #   aws-cdk-lib
contourpy==1.0.6
    # via matplotlib
cycler==0.11.0
    # via matplotlib
exceptiongroup==1.0.4
    # via cattrs
fonttools==4.38.0
    # via matplotlib
gitdb==4.0.10
    # via gitpython
gitpython==3.1.30
    # via -r requirements.in
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
joblib==1.2.0
    # via -r requirements.in
jsii==1.72.0
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-kubectl-v20
    #   aws-cdk-asset-node-proxy-agent-v5
    #   aws-cdk-aws-lambda-python-alpha
    #   aws-cdk-lib
    #   constructs
kiwisolver==1.4.4
    # via matplotlib
matplotlib==3.6.2
    # via -r requirements.in
numpy==1.23.5
    # via
    #   -r requirements.in
    #   contourpy
    #   matplotlib
    #   pandas
    #   pyarrow
    #   scipy
packaging==22.0
    # via
    #   build
    #   matplotlib
pandas==1.5.2
    # via -r requirements.in
pep517==0.13.0
    # via build
pillow==9.3.0
    # via matplotlib
publication==0.0.3
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-kubectl-v20
    #   aws-cdk-asset-node-proxy-agent-v5
    #   aws-cdk-aws-lambda-python-alpha
    #   aws-cdk-lib
    #   constructs
    #   jsii
pyarrow==10.0.1
    # via -r requirements.in
pyparsing==3.0.9
    # via matplotlib
python-dateutil==2.8.2
    # via
    #   botocore
    #   jsii
    #   matplotlib
    #   pandas
pytz==2022.6
    # via pandas
s3transfer==0.6.0
    # via boto3
scipy==1.9.3
    # via -r requirements.in
six==1.16.0
    # via python-dateutil
smmap==5.0.0
    # via gitdb
toml==0.10.2
    # via -r requirements.in
tomli==2.0.1
    # via
    #   build
    #   pep517
tqdm==4.64.1
    # via -r requirements.in
typeguard==2.13.3
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-kubectl-v20
    #   aws-cdk-asset-node-proxy-agent-v5
    #   aws-cdk-aws-lambda-python-alpha
    #   aws-cdk-lib
    #   constructs
    #   jsii
types-awscrt==0.16.1
    # via
    #   botocore-stubs
    #   types-s3transfer
types-boto3==1.0.2
    # via -r requirements.in
types-botocore==1.0.2
    # via -r requirements.in
types-pillow==9.3.0.4
    # via -r requirements.in
types-s3transfer==0.6.0.post5
    # via boto3-stubs
types-toml==0.10.8.1
    # via -r requirements.in
typing-extensions==4.4.0
    # via
    #   boto3-stubs
    #   jsii
urllib3==1.26.13
    # via botocore
